Add files via upload
This C++ program compares two integers, `a` and `b`, using the conditional (ternary) operator to determine the greater of the two. It then prints the greater number. The comparison is done with the expression `(a < b) ? b : a`, where if `a` is less than `b`, `b` is chosen; otherwise, `a` is chosen. The output displays the greatest number.