Add files via upload
This C++ program demonstrates exception handling. It defines a divide() function that throws an exception if the denominator is zero. In the main() function, the user is prompted to input a numerator and denominator. The program attempts to perform division and catches the exception if division by zero occurs, displaying an error message.